Refactor sagas.js to sagas/index.js
This refactoring puts all of the core sagas together in `src/sagas`
and will allow future refactoring to further break down the new
`src/sagas/index.js` to be a much easier and encapsulated task.
